Move configuration to chipmunk/config.rb
This allows chipmunk.rb to be responsible for dependencies, while the
config can be solely responsible for making the library configurable.
This is useful for test isolation; prior to this change, tests could
only be run by running the entire suite.

Add Chipmunk.root
