+++ spec/spec_helper.rb
@@ -5,6 +5,9 @@ require "faker"
 require "simplecov"
 SimpleCov.start
 
+require "chipmunk/config"
+Chipmunk.env = "test"
+
 RSpec.configure do |config|
   # rspec-expectations config goes here. You can use an alternate
   # assertion/expectation library such as wrong or the stdlib/minitest
